+++ factory-reset.py
@@ -1,6 +1,6 @@
 
-from proto97.usb import usb
-from proto97.sensor import factory_reset
+from proto9x.usb import usb
+from proto9x.sensor import factory_reset
 
 usb.open()
 factory_reset()
