+++ proto9x/init_flash.py
