+++ proto9x/util.py
